requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
elevenlabs==0.2.0
exceptiongroup==1.3.0
fastapi==0.115.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
gotrue==2.12.2
greenlet==3.2.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-openai==0.2.14
langchain-text-splitters==0.3.8
langchain-xai==0.1.0
langsmith==0.4.2
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.5.1
mypy_extensions==1.1.0
networkx==3.2.1
numpy==2.0.2
openai==1.92.2
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pluggy==1.6.0
postgrest==0.17.2
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
realtime==2.5.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.38.6
storage3==0.8.2
StrEnum==0.4.15
supabase==2.9.0
supafunc==0.6.2
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tomli==2.2.1
torch==2.7.1
tqdm==4.67.1
twilio==9.3.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.30.6
websockets==15.0.1
yarl==1.20.1
zstandard==0.23.0
pymongo==4.6.0
spacy==3.7.2
streamlit==1.28.0
plotly==5.17.0
sentence-transformers==2.2.2
pyannote-audio==3.1.0
ffmpeg-python==0.2.0
scikit-learn==1.3.0
xgboost==2.0.0